removes Constants.{cpp,hpp} and moves relevant part into SolverInterface.{cpp,hpp}
