+++ src/precice/SolverInterface.cpp
@@ -1,5 +1,7 @@
 #include "precice/SolverInterface.hpp"
 #include "precice/impl/SolverInterfaceImpl.hpp"
+#include "cplscheme/Constants.hpp"
+#include "io/Constants.hpp"
 
 namespace precice {
 
@@ -329,5 +331,40 @@ MeshHandle SolverInterface:: getMeshHandle
 {
   return _impl->getMeshHandle ( meshName );
 }
+
+namespace constants {
+
+const std::string& actionWriteInitialData()
+{
+  return cplscheme::constants::actionWriteInitialData();
+}
+
+const std::string& actionWriteIterationCheckpoint()
+{
+  return cplscheme::constants::actionWriteIterationCheckpoint();
+}
+
+const std::string& actionReadIterationCheckpoint()
+{
+  return cplscheme::constants::actionReadIterationCheckpoint();
+}
+
+const std::string& actionPlotOutput()
+{
+  static std::string plotOutput("plot-output");
+  return plotOutput;
+}
+
+int exportVTK()
+{
+  return io::constants::exportVTK();
+}
+
+int exportAll()
+{
+  return io::constants::exportAll();
+}
+
+} // namespace precice, constants
 
 } // namespace precice
