+++ src/sources.cmake
@@ -212,13 +212,10 @@ target_sources(precice
     src/partition/ReceivedPartition.cpp
     src/partition/ReceivedPartition.hpp
     src/partition/SharedPointer.hpp
-    src/precice/Constants.cpp
-    src/precice/Constants.hpp
     src/precice/MeshHandle.cpp
     src/precice/MeshHandle.hpp
     src/precice/SolverInterface.cpp
     src/precice/SolverInterface.hpp
-    src/precice/bindings/c/Constants.cpp
     src/precice/bindings/c/SolverInterfaceC.cpp
     src/precice/bindings/fortran/ConstantsFortran.cpp
     src/precice/bindings/fortran/ConstantsFortran.hpp
@@ -298,6 +295,7 @@ target_sources(precice
     src/utils/stacktrace.cpp
     src/utils/stacktrace.hpp
     src/versions.hpp
+    src/versions.hpp
     src/xml/ConfigParser.cpp
     src/xml/ConfigParser.hpp
     src/xml/XMLAttribute.hpp
@@ -310,9 +308,7 @@ target_sources(precice
 #
 
 set_property(TARGET precice PROPERTY PUBLIC_HEADER
-    src/precice/Constants.hpp
     src/precice/MeshHandle.hpp
     src/precice/SolverInterface.hpp
-    src/precice/bindings/c/Constants.h
     src/precice/bindings/c/SolverInterfaceC.h
     )
